Convert to typescript & RoPro instant join
