--- build_windows.bat
@@ -1,2 +1,0 @@
-@echo "[Win] Building"
-call electron-packager . rpresence --platform win32 --out ./build --overwrite --icon ./ico/logo.ico
