--- gamesconfig.json
@@ -1,49 +1,0 @@
-{
-  "version": "3.0.0",
-  "games": {
-    "13822889":{"name":"Lumber Tycoon 2","by":"Defaultio","iconkey":"13822889"},
-    "69184822":{"name":"Theme Park Tycoon 2","by":"Den_S","iconkey":"69184822"},
-    "132414630":{"name":"Speed Race","by":"MasterDaniel","iconkey":"132414630"},
-    "137877687":{"name":"RoCitizens","by":"Firebrand1","iconkey":"137877687"},
-    "137885680":{"name":"Zombie Rush","by":"Beacon Studio","iconkey":"137885680"},
-    "155615604":{"name":"Prison Life","by":"Aesthetical","iconkey":"155615604"},
-    "183364845":{"name":"Speed Run 4","by":"Vurse","iconkey":"183364845"},
-    "185655149":{"name":"Welcome to Bloxburg","by":"Coeptus","iconkey":"185655149"},
-    "205224386":{"name":"Hide and Seek Extreme","by":"Tim7775","iconkey":"205224386"},
-    "277751860":{"name":"Epic Minigames","by":"TypicalType","iconkey":"277751860"},
-    "286090429":{"name":"Arsenal","by":"ROLVe","iconkey":"286090429"},
-    "301549746":{"name":"Counter Blox","by":"ROLVe","iconkey":"301549746"},
-    "370731277":{"name":"MeepCity","by":"alexnewtron","iconkey":"370731277"},
-    "413424176":{"name":"Clone Tycoon 2","by":"Ultraw","iconkey":"413424176"},
-    "443406476":{"name":"Project Lazarus: ZOMBIES","by":"logitech101","iconkey":"443406476"},
-    "445664957":{"name":"Parkour","by":"hudzell","iconkey":"445664957"},
-    "447452406":{"name":"Robloxian Highschool","by":"Red Manta Studio","iconkey":"447452406"},
-    "488667523":{"name":"Those Who Remain","by":"Peak","iconkey":"488667523"},
-    "536102540":{"name":"Dragon Ball Z Final Stand","by":"SnakeWorl","iconkey":"536102540"},
-    "537413528":{"name":"Build A Boat For Treasure","by":"Chillz Studios","iconkey":"537413528"},
-    "606849621":{"name":"Jailbreak","by":"Badimo","iconkey":"606849621"},
-    "648362523":{"name":"Breaking Point","by":"Paradox","iconkey":"648362523"},
-    "734159876":{"name":"SharkBite","by":"Abracadabra","iconkey":"734159876"},
-    "735030788":{"name":"Royale High","by":"callmehbob","iconkey":"735030788"},
-    "855499080":{"name":"SKYWARS","by":"16bitplay Games","iconkey":"855499080"},
-    "893973440":{"name":"Flee the Facility","by":"A.W. Apps","iconkey":"893973440"},
-    "947714452":{"name":"Bakers Valley","by":"RoyStanford","iconkey":"947714452"},
-    "1029063110":{"name":"Bus Stop Simulator","by":"Steven_Scripts","iconkey":"1029063110"},
-    "1224212277":{"name":"Mad City","by":"Schwifty Studios","iconkey":"1224212277"},
-    "1262182609":{"name":"BOOGA BOOGA","by":"Soybeen","iconkey":"1262182609"},
-    "1417427737":{"name":"Mining Simulator","by":"Rumble Studios","iconkey":"1417427737"},
-    "1466995005":{"name":"Ragdoll Mayhem!","by":"LSPLASH","iconkey":"1466995005"},
-    "1537690962":{"name":"Bee Swarm Simulator","by":"Onett","iconkey":"1537690962"},
-    "1599679393":{"name":"Pet Simulator","by":"BIG Games","iconkey":"1599679393"},
-    "1672303943":{"name":"Ro-Ghoul","by":"Trapball","iconkey":"1672303943"},
-    "1962086868":{"name":"Tower of Hell","by":"YXCeptional Studios","iconkey":"1962086868"},
-    "2091563160":{"name":"Bus Simulator","by":"NobleReign","iconkey":"2091563160"},
-    "2346834658":{"name":"Weight Lifting Simulator 3","by":"Flamin' Studios","iconkey":"2346834658"},
-    "2377868063":{"name":"Strucid","by":"Frosted Studio","iconkey":"2377868063"},
-    "2414851778":{"name":"Dungeon Quest","by":"vCaffy","iconkey":"2414851778"},
-    "2551991523":{"name":"Broken Bones IV","by":"Zaquille","iconkey":"2551991523"},
-    "3173458677":{"name":"Draw It!","by":"litten's Games","iconkey":"3173458677"},
-    "2409854691":{"name":"BloxParty","by":"DareDev","iconkey":"2409854691"},
-    "738339342":{"name":"Flood Escape 2","by":"Crazyblox Games","iconkey":"738339342"}
-  }
-}
