+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "outDir": "build",
+        "noImplicitAny": false,
+        "removeComments": true,
+        "preserveConstEnums": true,
+	    "skipLibCheck": true,
+        "esModuleInterop": true
+    },
+    "include": ["src"],
+    "exclude": ["node_modules", "**/*.spec.ts"]
+}
